Ejercicio 9.5

+++ data/src/main/java/es/danirofe/data/FilmDto.kt
@@ -6,6 +6,6 @@ data class FilmDto(
     @SerializedName("id") val id:Int,
     @SerializedName("title") val title:String,
     @SerializedName("overview") val description: String,
-    @SerializedName("vote_average") val rating: Double,
+    @SerializedName("vote_average") val rating: Float,
     @SerializedName("poster_path") val imageUrl:String
 )
